+++ src/models.rs
@@ -525,7 +525,6 @@ pub trait FromDirective {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
     use crate::{parser::Tokenizer, CommonOpts};
 
     #[test]